feat!: `REPLStart` now always creates a new REPL when count = 0.
Previously, if count was set to 0, `REPLStart` would attempt to create a
REPL with ID = 1. If a REPL with ID = 1 already existed, the command
would simply focus on the existing instance instead of creating a new
one.

With this update, when count is 0, the command will always create a new
REPL with an incrementally assigned ID.